Safari-611.1.10/LayoutTests/platform/gtk/http/wpt/beacon/beacon-async-error-logging-expected.txt

CONSOLE MESSAGE: Beacon API cannot load http://invalid.localhost/. Error resolving “invalid.localhost”: Name or service not known

PASS Should log an error message in the console
